server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.2.6
pytest==8.2.2
pydantic-settings==2.4.0
python-multipart==0.0.6

# Deepfake Detection & Generation Libraries
deepface==0.0.95
face-recognition==1.3.0
dlib==20.0.0
tensorflow==2.20.0
tf-keras==2.20.1
keras==3.12.0
mtcnn==1.0.0
retina-face==0.0.17

# InsightFace & Advanced Face Analysis
insightface==0.7.3
onnxruntime==1.23.2
onnx==1.19.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
